streamlit/charging_analysis/requirements.txt
# 核心依赖
reportlab==4.3.1
streamlit==1.33.0
pandas==2.2.1
numpy==1.26.4
pandasai>=2.0.4  # 2.0.4+支持pandas 2.x
altair==5.3.0

# 数据可视化
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.18.0

# 数据处理
scipy==1.12.0
scikit-learn==1.4.0

# 其他工具
python-dotenv==1.0.1
openpyxl==3.1.2
